backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.122.0",
    "geoalchemy2>=0.18.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-schemadisplay>=2.0",
    "sqlalchemy-utils>=0.42.1",
    "datascience",
    "pyjwt>=2.10.1",
]

[tool.pytest.ini_options]
# Add the 'backend/' directory to the Python path so tests can find src/ etc.
pythonpath = [
    ".",
]
testpaths = [
    "tests"
]
asyncio_mode = "auto"

[tool.ruff]
# Ignore unused imports (F401) in all SQLAlchemy model files.
lint.per-file-ignores = { "src/models/*.py" = ["F401"] }
exclude = [
    "alembic/versions/*",
]

[tool.uv.sources]
datascience = { path = "../datascience", editable = true }
